+++ src/hooks/useThrottling.ts
@@ -1,13 +1,17 @@
 import { useRef } from "react";
 
-export const useThrottling = <TEvent>(fn: (event: TEvent) => void, limit: number = 8) => {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const useThrottling = <T extends (...args: any[]) => void>(
+  fn: T,
+  limit: number = 8
+): [(...args: Parameters<T>) => void] => {
   const lastUpdateTime = useRef<number>(0);
 
-  const throttledFunction = (event: TEvent) => {
+  const throttledFunction = (...args: Parameters<T>) => {
     const now = Date.now();
 
     if (now - lastUpdateTime.current >= limit) {
-      fn(event);
+      fn(...args);
       lastUpdateTime.current = now;
     }
   };